requirements.txt
appdirs==1.4.3
attrs==18.2.0
lark-parser==0.6.6
mccabe==0.6.1
mpmath==1.1.0
nose==1.3.7
numeral==0.1.0.11
numpy==1.15.4
parso==0.3.1
pkg-resources==0.0.0
pycodestyle==2.4.0
pyflakes==2.0.0
recordclass==0.8.3
rope==0.11.0
scipy==1.2.0
snakeviz==1.0.0
sympy==1.3
toml==0.10.0
tornado==5.1.1
yapf==0.26.0
